telegram client: fix infinite recursion in send_file
The iterability check added in 36e21019
does not tell away strings from containers
of strings, therefore all possible arguments
are considered iterable, and the function
recurses into itself with each of the items,
which again will be considered iterable...

So doing now an explicit type check for str
before getting to itemizing. (XXX I wonder if
bytes/bytearray objects could be taken as
arguments? If yes, they should as well be
included in the type check.)