Improve Quart example
As the client seems tied to a specific login, I've changed it from
being a global to a local in the request handler. This should prevent
one user logging in and another unrelated user having access (due to a
global shared client). I've then made use of session storage to store
the phone and code (session storage is per user and deleted when the
browser tab is closed). I wasn't sure if the code was sensitive, if so
it is potentially problematic to store in the default session system
(cookies) - a server storage solution (e.g. redis or in memory) would
be better.

Note that if the client has to be global the Quart before/after
serving functions should be used to connect and disconnect, as so,

    client = ...

    @app.before_serving
    async def startup():
        await client.connect()

    @app.after_serving
    async def cleanup():
        await client.disconnect()

as this creates the client within the event-loop managed by
Quart. Alternatively if the event-loop must be managed outside of
Quart the following should be used,

    from hypercorn.asyncio import serve

    loop.run_until_complete(serve(app))

instead of `app.run()`.

I've then made a number of more minor changes to use the template
rendering system within Quart (more features, better example) and
to clean up the code.