setup.py: Make build reproducible by sorting generator input
Debian's reproducible builds infrastructure reports that the ordering of
objects listed in the tlobjects dict in
/usr/lib/python3/dist-packages/telethon/tl/alltlobjects.py
depends on the filesystem ordering of the *.tl files read at build time.

Fix that by sorting the list in setup.py.

Reproducible builds are desirable for security and auditing reasons, and
easing QA. More details at https://reproducible-builds.org/ .